Reorganized home components
