+++ src/pages/home/components/countries/index.js
@@ -1,0 +1,3 @@
+import Component from './component';
+
+export default Component;
