+++ src/pages/home/components/countries/styles.scss
@@ -1,0 +1,25 @@
+.c-choose-country {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+
+  .country {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    align-items: center;
+
+    img {
+      transition: all 0.1s cubic-bezier(.445, .05, .55, .95); // $ease-in-out-sine
+      transform-origin: left;
+    }
+
+    &:hover {
+      img {
+        transform: scale(1.4);
+      }
+    }
+  }
+}
