+++ src/pages/home/styles.scss
@@ -1,7 +1,7 @@
 @import '~styles/settings';
 
 .c-home {
-  background-image: url('/background.png');
+  background-image: url('/background.svg');
   background-repeat: no-repeat;
   background-attachment: fixed;
   background-position: right center;
@@ -24,32 +24,6 @@
       font-size: 18px;
       line-height: 26px;
       max-width: 700px;
-    }
-  }
-}
-
-.c-choose-country {
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  .country {
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-    align-items: center;
-
-    img {
-      transition: all 0.1s cubic-bezier(.445, .05, .55, .95); // $ease-in-out-sine
-      transform-origin: left;
-    }
-
-    &:hover {
-      img {
-        transform: scale(1.4);
-      }
     }
   }
 }
